SmoothedResultManager now reduced to one.
