+++ rifServices/src/main/resources/RIFLogger.properties
@@ -101,7 +101,6 @@ rifServices.dataStorageLayer.common.CommonStudyStateManager=true
 rifServices.dataStorageLayer.pg.PGSQLStudyExtractManager=true
 rifServices.dataStorageLayer.pg.PGSQLMapDataManager=true
 rifServices.dataStorageLayer.pg.PGSQLAgeGenderYearManager=true
-rifServices.dataStorageLayer.pg.PGSQLSmoothedResultManager=true
 
 rifGenericLibrary.dataStorageLayer.ms.MSSQLQueryUtility=true
 rifGenericLibrary.dataStorageLayer.ms.MSUserDatabaseConnections=true
@@ -120,7 +119,7 @@ rifServices.dataStorageLayer.common.RunStudyThread=true
 rifServices.dataStorageLayer.common.CommonHealthOutcomeManager=true
 rifServices.dataStorageLayer.ms.MSSQLStudyExtractManager=true
 rifServices.dataStorageLayer.ms.MSSQLMapDataManager=true
-rifServices.dataStorageLayer.ms.MSSQLSmoothedResultManager=true
+rifServices.dataStorageLayer.common.SmoothedResultManager=true
 rifServices.dataStorageLayer.common.CommonUserService=true
 
 rifServices.businessConceptLayer.Investigation=true
